Move some metadata related tests into GifMetaDataTests.cs
